Delete `test/unit.js`
